Rebase core/opl and cleanup
- rebase circular import issue fix to core/opl
- remove stale unused imports
