+++ tests/test_retry.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import unittest
-import argparse
 import datetime
 
 from .context import opl
